[feat] Add new alerter system (info, success and error alerts)
