--- src/components/Notifier.jsx
@@ -1,111 +1,0 @@
-/** @jsx h */
-import { h, Component } from 'preact'
-
-const createStore = () => {
-  let notifications = []
-  let listeners = []
-
-  const dispatch = (notification) => {
-    notification.id = notifications.length
-    notifications.push(notification)
-    listeners.forEach(listener => listener(notification))
-  }
-
-  const subscribe = (listener) => {
-    listeners.push(listener)
-  }
-
-  return { dispatch, subscribe }
-}
-
-const store = createStore()
-
-class Notification extends Component {
-  componentDidMount () {
-    this.close = this.close.bind(this)
-    this.closeTimer = setTimeout(() => {
-      this.beginClosing()
-    }, 2000)
-  }
-
-  componentWillUnmount () {
-    this.base.removeEventListener('transitionend', this.close, false)
-    this.base.classList.remove('coz-alert--closing')
-    if (this.closeTimer) {
-      clearTimeout(this.closeTimer)
-    }
-  }
-
-  beginClosing () {
-    this.base.addEventListener('transitionend', this.close, false)
-    this.base.classList.add('coz-alert--closing')
-  }
-
-  close () {
-    this.props.onClose()
-  }
-
-  render () {
-    const { message, type } = this.props
-    return (
-      <div className={['coz-alert', 'coz-alert--' + type].join(' ')}>
-        <div className='coz-alert-title'>{message}</div>
-      </div>
-    )
-  }
-}
-
-export default class Notifier extends Component {
-  constructor (props, context) {
-    super(props, context)
-    this.state = {
-      notifications: []
-    }
-  }
-
-  static info (msg) {
-    store.dispatch({ type: 'info', msg })
-  }
-
-  static warning (msg) {
-    store.dispatch({ type: 'warning', msg })
-  }
-
-  static error (msg) {
-    store.dispatch({ type: 'error', msg })
-  }
-
-  componentDidMount () {
-    store.subscribe(this.notify.bind(this))
-  }
-
-  notify (notification) {
-    this.setState({
-      notifications: [...this.state.notifications, notification]
-    })
-  }
-
-  handleClose (id) {
-    let idx = this.state.notifications.findIndex(n => n.id === id)
-    this.setState({
-      notifications: [
-        ...this.state.notifications.slice(0, idx),
-        ...this.state.notifications.slice(idx + 1)
-      ]
-    })
-  }
-
-  render () {
-    return (
-      <div className='coz-alerter'>
-        {this.state.notifications.map(notif => (
-          <Notification
-            type={notif.type}
-            message={notif.msg}
-            onClose={this.handleClose.bind(this, notif.id)}
-          />
-        ))}
-      </div>
-    )
-  }
-}
