+++ src/styles/alerter.styl
@@ -1,0 +1,12 @@
+@import 'cozy-ui'
+
+.set-alerter
+    @extends $alerts
+
+    .coz-overlay
+        // Specify z-index to bypass upload button z-index
+        z-index 59
+
+    .coz-alert
+        // Specify z-index to bypass upload button z-index
+        z-index 60
