+++ src/styles/main.styl
@@ -1,6 +1,2 @@
 @import 'cozy-ui/defaults'
 @import 'cozy-ui'
-
-:global
-    // Global utils classes comes here
-    @extends $alerts
